requirements.txt
dominate==2.6.0
imageio
ipython
lpips==0.1.3
matplotlib
numpy
opencv-python
pandas
Pillow
scikit-image
tensorboard
tensorboardX
tifffile==2020.9.3
torch==1.1.0
torchfile==0.1.0
torchvision==0.3.0
tqdm
